# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/order/create": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create A New Admin Account",
                "operationId": "785adac124bc0364d40a53e50567f602",
                "requestBody": {
                    "description": "Supply Order Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderProperties"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/order/{uuid}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show Order",
                "operationId": "69de54d0d3b16e2105704e85dd5b9fcb",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of Order",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "uuiuytyytytj5656jk-jnnknnkbkjnk-nghn6n565"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update Order",
                "operationId": "b994d7152c7d94e19753d17ba55082d8",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of Order",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "uuiuytyytytj5656jk-jnnknnkbkjnk-nghn6n565"
                    }
                ],
                "requestBody": {
                    "description": "Supply Order Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderProperties"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an Order",
                "operationId": "a6290d6e07443d8b92ecf8283f142d6a",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of Order",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "uuiuytyytytj5656jk-jnnknnkbkjnk-nghn6n565"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/order": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List All Orders",
                "operationId": "a664bc10b36e2203fba446103026ad1b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page Number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Item Per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort By",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Direction of sort",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/order/{uuid}/download": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Downlad an Order",
                "operationId": "1648a871a2e4e7650baafe047c8cb82e",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of Order",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "uuiuytyytytj5656jk-jnnknnkbkjnk-nghn6n565"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateOrderProperties": {
                "required": [
                    "order_status_uuid",
                    "products",
                    "payment_uuid",
                    "address"
                ],
                "properties": {
                    "order_status_uuid": {
                        "type": "string",
                        "format": "Order Status UUID",
                        "example": ""
                    },
                    "payment_uuid": {
                        "type": "string",
                        "format": "Payment UUID",
                        "example": ""
                    },
                    "products": {
                        "type": "array",
                        "format": "Array of objects with product uuid and quantity",
                        "items": {
                            "type": "object"
                        },
                        "example": "[]"
                    },
                    "address": {
                        "type": "object",
                        "format": "Billing and Shipping address",
                        "example": "{}"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CreateOrderProperties"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "scheme": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}